requirements.txt
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.0.3
PyJWT==2.1.0